refactor: inbox
